number should be casted with long instead of int
number should be casted with long instead of int

+++ main.cpp
@@ -33,7 +33,7 @@ namespace nlohmann
                 double number = j.get<double>();
                 if (number == std::floor(number))
                 {
-                    return py::int_(j.get<int>());
+                    return py::int_(j.get<long>());
                 }
                 else
                 {
